Renamed 'readytosend' package to 'send' and moved it to 'instancemanagemen'
