+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/external/InstanceUploadActionTest.kt
@@ -7,7 +7,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
-import org.odk.collect.android.readytosend.InstanceUploaderActivity
+import org.odk.collect.android.instancemanagement.send.InstanceUploaderActivity
 import org.odk.collect.android.support.pages.OkDialog
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.TestRuleChain
