+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyComponent.java
@@ -12,8 +12,8 @@ import org.odk.collect.android.activities.FormFillingActivity;
 import org.odk.collect.android.activities.FormHierarchyActivity;
 import org.odk.collect.android.activities.FormMapActivity;
 import org.odk.collect.android.activities.InstanceChooserList;
-import org.odk.collect.android.readytosend.InstanceUploaderActivity;
-import org.odk.collect.android.readytosend.InstanceUploaderListActivity;
+import org.odk.collect.android.instancemanagement.send.InstanceUploaderActivity;
+import org.odk.collect.android.instancemanagement.send.InstanceUploaderListActivity;
 import org.odk.collect.android.adapters.InstanceUploaderAdapter;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.application.initialization.ApplicationInitializer;
