+++ collect_app/src/main/java/org/odk/collect/android/mainmenu/MainMenuActivity.kt
@@ -26,9 +26,9 @@ import org.odk.collect.android.databinding.MainMenuBinding
 import org.odk.collect.android.formlists.blankformlist.BlankFormListActivity
 import org.odk.collect.android.formmanagement.FormFillingIntentFactory
 import org.odk.collect.android.injection.DaggerUtils
+import org.odk.collect.android.instancemanagement.send.InstanceUploaderListActivity
 import org.odk.collect.android.projects.ProjectIconView
 import org.odk.collect.android.projects.ProjectSettingsDialog
-import org.odk.collect.android.readytosend.InstanceUploaderListActivity
 import org.odk.collect.android.utilities.ApplicationConstants
 import org.odk.collect.android.utilities.ThemeUtils
 import org.odk.collect.androidshared.ui.DialogFragmentUtils.showIfNotShowing
